Move response data manipulation into dataAdaptors for word and phrase
- make use of `useParams` `sitename`
